feat: Comprehensive code audit and critical fixes
This commit delivers a comprehensive code audit and implements fixes for several critical issues identified during the review.

The audit covered code quality, security, performance, best practices, and testing. A detailed report has been created at `COMPREHENSIVE_AUDIT_REPORT.md`.

Key changes implemented in this commit include:

1.  **Configuration Cleanup:** Consolidated `next.config.js` and `next.config.ts` into a single `next.config.mjs` file, removing dangerous flags and improving maintainability.

2.  **Security Hardening:**
    *   Centralized the application of security headers, including a strict Content Security Policy (CSP), into the Next.js middleware.
    *   This resolves a critical vulnerability where the application was not protected against attacks like XSS and clickjacking.

3.  **Database Schema Refactoring:**
    *   Removed the legacy and orphaned `inquiry_deals` model from the Prisma schema, eliminating data model inconsistency and potential for future errors.

These changes significantly improve the security posture, maintainability, and data integrity of the application. The audit report contains further recommendations for critical issues that need to be addressed in the application code, such as an exposed admin API on the home page and business logic flaws in the deals router.
