--- next.config.ts
@@ -1,142 +1,0 @@
-import type { NextConfig } from 'next';
-import { securityHeadersConfig } from './src/lib/security-headers';
-
-const nextConfig: NextConfig = {
-  // Security headers
-  async headers() {
-    return [
-      {
-        source: '/(.*)',
-        headers: securityHeadersConfig,
-      },
-      {
-        source: '/api/(.*)',
-        headers: [
-          {
-            key: 'X-Content-Type-Options',
-            value: 'nosniff',
-          },
-          {
-            key: 'X-Frame-Options',
-            value: 'DENY',
-          },
-          {
-            key: 'X-XSS-Protection',
-            value: '1; mode=block',
-          },
-          {
-            key: 'Access-Control-Allow-Origin',
-            value: process.env.NODE_ENV === 'production' 
-              ? 'https://geodomainland.com' 
-              : 'http://localhost:3000',
-          },
-          {
-            key: 'Access-Control-Allow-Methods',
-            value: 'GET, POST, PUT, DELETE, OPTIONS',
-          },
-          {
-            key: 'Access-Control-Allow-Headers',
-            value: 'Content-Type, Authorization, X-Requested-With',
-          },
-          {
-            key: 'Access-Control-Allow-Credentials',
-            value: 'true',
-          },
-        ],
-      },
-      {
-        source: '/static/(.*)',
-        headers: [
-          {
-            key: 'Cache-Control',
-            value: 'public, max-age=31536000, immutable',
-          },
-        ],
-      },
-    ];
-  },
-
-  // Security configurations
-  poweredByHeader: false,
-  compress: true,
-
-  // Image optimization
-  images: {
-    domains: ['res.cloudinary.com', 'localhost'],
-    formats: ['image/webp', 'image/avif'],
-    deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
-    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
-    minimumCacheTTL: 60,
-    dangerouslyAllowSVG: false,
-    contentSecurityPolicy: "default-src 'self'; script-src 'none'; sandbox;",
-  },
-
-  // Experimental features for security
-  experimental: {
-    optimizeCss: true,
-    optimizePackageImports: ['@radix-ui/react-icons', 'lucide-react'],
-    turbo: {
-      rules: {
-        '*.svg': {
-          loaders: ['@svgr/webpack'],
-          as: '*.js',
-        },
-      },
-    },
-  },
-
-  // Webpack configuration for security
-  webpack: (config, { dev, isServer }) => {
-    // Security optimizations
-    if (!dev) {
-      config.optimization.minimize = true;
-    }
-
-    // Add security headers to webpack
-    config.plugins.push(
-      new (require('webpack')).DefinePlugin({
-        'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-      })
-    );
-
-    return config;
-  },
-
-  // Environment variables
-  env: {
-    CUSTOM_KEY: process.env.CUSTOM_KEY,
-  },
-
-  // Redirects for security
-  async redirects() {
-    return [
-      {
-        source: '/admin',
-        destination: '/admin/dashboard',
-        permanent: false,
-      },
-      {
-        source: '/api/auth/signin',
-        destination: '/login',
-        permanent: false,
-      },
-      {
-        source: '/api/auth/signout',
-        destination: '/logout',
-        permanent: false,
-      },
-    ];
-  },
-
-  // Rewrites for security
-  async rewrites() {
-    return [
-      {
-        source: '/api/health',
-        destination: '/api/health/check',
-      },
-    ];
-  },
-};
-
-export default nextConfig;
